% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_bigrams.R
\name{create_bigrams}
\alias{create_bigrams}
\title{Create bigrams from text}
\usage{
create_bigrams(input, text_col)
}
\arguments{
\item{input}{A data frame containing text data}

\item{text_col}{Name of the text column}
}
\value{
A tibble containing bigrams
}
\description{
Convert a text column into 2 word combinations
}
\examples{
test_input <- data.frame(
  text = c(
    "Lisan Al Gaib leads the way",
    "Messi is the Goat"
  )
)

create_bigrams(test_input, "text")

}
