% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_bigrams.R
\name{plot_top_bigrams}
\alias{plot_top_bigrams}
\title{Plot the most common bigrams}
\usage{
plot_top_bigrams(bigram_input, n = 10)
}
\arguments{
\item{bigram_input}{A data frame containing bigrams}

\item{n}{Number of top bigrams to plot}
}
\value{
A ggplot bar chart.
}
\description{
Create a bar chart of the most frequent bigrams
}
\examples{
test_input <- data.frame(
  text = c(
    "Lisan Al Gaib leads the way",
    "Messi is the Goat"
  )
)

bigrams <- create_bigrams(test_input, "text")

plot_top_bigrams(bigrams)

}
