% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_text.R
\docType{data}
\name{sample_text}
\alias{sample_text}
\title{Sample text dataset}
\format{
A data frame with 2 rows and 2 variables:
\describe{
\item{id}{Text ID}
\item{text}{Example text paragraph}
}
}
\source{
Created for package examples.
}
\usage{
sample_text
}
\description{
Example text data for tokenization and bigram analysis/ exploration
}
\keyword{datasets}
