% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tokenize_words.R
\name{tokenize_words}
\alias{tokenize_words}
\title{Tokenize text into words}
\usage{
tokenize_words(input, text_col)
}
\arguments{
\item{input}{A data frame containing text data}

\item{text_col}{Name of the text column}
}
\value{
A tibble with one word per row
}
\description{
Convert a text column into individual word tokens
}
\examples{
test_input <- data.frame(
  text = c(
    "Lisan Al Gaib leads the way",
    "Messi is the Goat"
  )
)

tokenize_words(test_input, "text")

}
